+++ ROWM/Controllers/ExpenseController.cs
@@ -245,7 +245,7 @@ namespace ROWM.Controllers
 
         public string ExpenseTitle { get; set; }
 
-        public double ExpenseAmount { get; set; }
+        public decimal ExpenseAmount { get; set; }
 
         public string Notes { get; set; }
